Remove ngraph_builder class and node_context_impl class
