+++ ngraph/frontend/tensorflow/src/op/max_pool.cpp
@@ -48,7 +48,7 @@ OutputVector TranslateMaxPoolOp(const NodeContext& node) {
 
     CoordinateDiff padding_below;
     CoordinateDiff padding_above;
-    Builder::MakePadding(tf_padding_type,
+    MakePadding(tf_padding_type,
                          ng_image_shape,
                          ng_kernel_shape,
                          ng_strides,
