+++ ngraph/frontend/tensorflow/src/op/unary_op.cpp
@@ -41,7 +41,7 @@ OutputVector TranslateUnaryOp(const NodeContext& op, std::function<Output<Node>(
     Output<Node> ng_input = op.get_ng_input(0);
     auto ng_node = create_unary_op(ng_input);
     if (ng_node != ng_input) {
-        Builder::SetTracingInfo(op.get_name(), ng_node);
+        SetTracingInfo(op.get_name(), ng_node);
     }
     // SaveNgOp(ng_op_map, node.get_name(), ng_node);
     // return Status::OK();
